+++ geo-traits/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- BREAKING: Mark `CoordTrait::nth_unchecked` as `unsafe` and add `CoordTrait::nth_or_panic`.
+  - <https://github.com/georust/geo/pull/1242>
+
 ## 0.1.1
 
 - Fix `TriangleTrait::second` and `TriangleTrait::third` to return the second and third coordinates instead of the first.